Add menu.py file and move menu into file, adapt code for the new situation
